✨ Feat: Add Code To Install ZSH Extensions [main]
Added some in the install_zshrc() function, to install oh-my-zsh and
some extensions to be used in the .zshrc file.

Changed backup process so that the bin directory is copied instead of
being moved to keep the scripts that were previously in the directory to
be preserved for future use and also added -u argument to rsync to
update the bin directory.

Changes Made:
	- Added code to install zsh addons. (install_dotfiles.sh)
	- Changed mv command to cp -r in bin directory backup.
	  (installdotfiles.sh)
	- Added -u argument in bin directory rsync. (installdotfiles.sh)
	- Added comment above install_utility_configs() function.
	  (installdotfiles.sh)

Files Changed:
	modified:   installdotfiles.sh

Date: 07/05/2023, 12:18 P.M. IST

All Done.