✨ Feat: Add Execution Part Of Script [main]
Added the execution part of the script to call all the functions defined
and install the dotfiles. It also recommends users to file issues and
make pull requests.

Add a check_requirements() function to check if the required commands
are installed or not.

Fixed filename typos in the syncing commands of topgrade and starship.

Changed the fill layout of the starship prompt to show dashes with
spaces.

Changes Made:
	- Added calling of functions in execution part. (installdotfiles.sh)
	- Added check_requirements function. (installdotfiles.sh)
	- Changed fill layout in starship prompt. (starship.toml)
	- Fixed filename typos. (installdotfiles.sh)

Files Changed:
	modified:   installdotfiles.sh
	modified:   starship.toml

Date: 09/05/2023, 12:23 P.M. IST

All Done.