🔥 Removal: Remove Shell Redirects [main]
Removed shell redirects from the dotfiles installation scripts as it was
causing inputs not working as found during a test. It will be added
when a better solution is found

Removed oh-my-zsh installation code because it just installed it and did
not let the script continue afterwards instead made it requirement for
the script.

Added code to create terminal-utilities directory before the
installation starts in the script so the zsh extension can be cloned
in the directory.

Changes Made:
	- Removed shell redirects for logging. (installdotfiles.sh)
	- Removed oh-my-zsh installation code. (installdotfiles.sh)
	- Added mkdir command to create terminal utilities directory.
	  (installdotfiles.sh)

Files Changed:
	modified:   installdotfiles.sh

Date: 18/05/2023, 10:31 A.M. IST

All Done.