🐛 BugFix: Fix Utilities Installation Bug [main]
External utilities with config files in their separate directory in the
.config directory didn't get installed sometimes if the separate
directory didn't already exist. Now before syncing or backing up the
config file, the script first checks for the existence of the directory
and if it does not exist it creates it first and then syncs/backs up the
file.

Added oh-my-zsh command as a requirement to run the script.

Changes Made:
	- Added check for the existence of the config file directories
	  of external utilities. (installdotfiles.sh)
	- Added omz as a requirement. (installdotfiles.sh)

Files Changed:
	modified:   installdotfiles.sh

Date: 19/05/2023, 11:18 A.M. IST

All Done.