✨ Feat: Add Dotfiles Selection Menu [main]
Added a new and improved method of installing and syncing dotfiles which
makes the process more flexible and modular. The process is now
separated into 3 files
1. sync-functions.sh: It contains all the functions to install dotfiles.
2. install.sh: It contains the menu and does the installation process.
3. dotfiles.sh: This file is created by install.sh to call the functions
of sync-functions.sh by sourcing the file and executing only selected
functions from the menu.
This approach makes it easier to spot bugs, check requirements for the
dotfiles and the script. It also makes the process more customizable.

Changes Made:
	- Added a installation menu. (install.sh)
	- Added file containing installation functions. (sync-functions.sh)
	- Removed the position configuration for NERDTRee in vim. (vimrc)

Files Changed:
	modified:   .gitignore
	new file:   install.sh
	deleted:    installdotfiles.sh
	new file:   sync-functions.sh
	modified:   vimrc

Date: 14/01/2025, 12:37 A.M. IST

All Done.