🚸 Usability: Update Files For Better UX [main]
Changes Made:
	- Added aliases for zellij commands.
	- Removed brash binary file since it can be replaced by trash-cli.
	- Added the gnome-extensions file in the repository for ease of use.
	- Added the .local/share cargo bin folder to the PATH.
	- Changed dependency for zsh and gnome extensions syncing.
	- Changed superfile editor and syntax highlight.
	- Added copy-on-select in zellij.
	- Added autostart for zellij in zsh.

Added aliases and autostart for zellij in the zsh session to make it
easier to use. Removed brash from bin folder as it can replaced by
trash-cli which is far better.

Added gnome-extensions and config file to the repo for convenience as
instead of being in the home directory it can be directly shipped with
the dotfiles, this is a much cleaner alternative to installing it
manually. Added another cargo bin folder to the PATH variable.

Changed the dependency for zsh syncing from omz (oh-my-zsh) to just
simple zsh and removed dconf as dependency for syncing the
gnome-extensions as it is preinstalled. Removed the changes made to the
superfile editor. Added copy-on-select setting to zellij.

Files Changed:
	modified:   aliases
	deleted:    bin/brash
	new file:   gnome-extensions
	new file:   gnome-extensions-config.ini
	modified:   home_env
	modified:   install.sh
	modified:   superfile_config.toml
	modified:   sync-functions.sh
	modified:   zellij_config.kdl
	modified:   zshrc

Date: 08/05/2025, 04:05 P.M. IST

All Done.

Signed-off-by: Kunal Kumar <kunalkumarchourasiya2021@gmail.com>